backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
python-decouple==3.8
psycopg2-binary==2.9.9
django-filter==23.5
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
stripe==7.8.0
boto3==1.34.0
django-storages==1.14.2
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
